+++ request_api_helper/lib/helper/database.dart
@@ -1,13 +1,14 @@
 import 'dart:io';
 
-import 'package:request_api_helper/helper.dart';
+import 'package:request_api_helper/helper/encrypt.dart';
 import 'package:sqflite/sqflite.dart';
 
 class DatabaseCompute {
   String? name;
   String? path;
+  Crypt? encrypt;
 
-  DatabaseCompute({this.name, this.path});
+  DatabaseCompute({this.name, this.path, this.encrypt});
 }
 
 class StorageBase {
@@ -19,6 +20,15 @@ class StorageBase {
       path = databasesPath + 'sharedPreferences.db';
     } catch (_) {
       print(_);
+    }
+  }
+
+  static _decrypt(Crypt? keys, text) {
+    try {
+      if (keys != null) return keys.decode(text);
+      return text;
+    } catch (_) {
+      return text;
     }
   }
 
@@ -46,7 +56,7 @@ class StorageBase {
   static deleteAll(DatabaseCompute data) async {
     Database db = await connect();
     await db.transaction((txn) async {
-      await txn.rawInsert('DELETE FROM data');
+      await txn.rawInsert('DELETE FROM data WHERE name!=_encrypted_session_master');
     });
   }
 
@@ -60,20 +70,19 @@ class StorageBase {
   static Future<String?> getString(DatabaseCompute data) async {
     Database db = await connect();
     List<Map> list = await db.rawQuery('SELECT * FROM data WHERE name="${data.name}" AND type="string"');
-
-    return list.isEmpty ? null : list.first['text'];
+    return list.isEmpty ? null : _decrypt(data.encrypt, list.first['text']);
   }
 
   static Future<double?> getDouble(DatabaseCompute data) async {
     Database db = await connect();
     List<Map> list = await db.rawQuery('SELECT * FROM data WHERE name="${data.name}" AND type="double"');
-    return list.isEmpty ? null : double.tryParse(list.first['text']);
+    return list.isEmpty ? null : double.tryParse(_decrypt(data.encrypt, list.first['text']));
   }
 
   static Future<int?> getInt(DatabaseCompute data) async {
     Database db = await connect();
     List<Map> list = await db.rawQuery('SELECT * FROM data WHERE name="${data.name}" AND type="integer"');
-    return list.isEmpty ? null : int.tryParse(list.first['text']);
+    return list.isEmpty ? null : int.tryParse(_decrypt(data.encrypt, list.first['text']));
   }
 
   static Future<bool?> getBool(DatabaseCompute data) async {